feat(config): format order of generated talconfig
By default, after file is converted to json to patch with jsonpatch,
all the keys will be sorted from A-Z, which looks so hard to read. This
will at least format it to be more reasonable. Although everybody
preferences are different, but at least we don't have weird order like
maps with a long list of nested maps/slices are at the top and simple
`key: value` map in the middle etc.