feat(genconfig): can generate config using talsecret.yaml
You can now put a `talsecret.yaml` file and `talhelper` will use that
file to generate config for you. The old behaviour of using
`talenv.yaml` will still be supported, but I think `talenv.yaml` is more
of for variable substitution while `talsecret.yaml` is for cluster
secret tokens and keys (it can only be used for that too).

You can migrate to the new `talsecret.yaml` like this:
1. `talhelper gen secret --from-configfile ./clusterconfig/mynode.yaml > talsecret.yaml`
2. Encrypt talsecret.yaml with SOPS.
3. Remove all the `inlinePatch` created by `gensecret --patch-configfile` flag
4. Optionally delete the `talenv.yaml` file.