feat(config): new patches key to replace configPatches and inlinePatch
Thanks to the guidance of frezbo from Talos and the code from
siderolabs, now you can use `patches` that can do everything the former
`configPatches` and `inlinePatch` can do. And not only that, you can now
also use `"@filepath"` to load a patch from a file. You can still use
variable substitution from in that file, making it a template patch.

The old `configPatches` and `inlinePatch` will still work just fine, but
the new `patches` is so much nicer. Here's a comparison on how they
differ:

***Old `configPatches` and `inlinePatch`***
```
`talconfig.yaml`
controlPlane:
  configPatches:
    - op: add
      path: /machine/kubelet/extraArgs
      value:
        rotate-server-certificate: "true"
    - op: add
      path: /machine/install/extraKernelArgs
      value:
        - talos.logging.kernel=${udpAddress}
  inlinePatch:
    machine:
      network:
        interfaces:
          - interface: eth1
            addresses: [192.168.200.11/24]
```

***New `patches`***
`talconfig.yaml`:
```
controlPlane:
  patches:
    - |-
      - op: add
        path: /machine/kubelet/extraArgs
        value:
          rotate-server-certificate: "true"
    - "@patch-file.yaml"
    - |-
      machine:
        network:
          interfaces:
            - interface: eth1
              addresses: [192.168.200.11/24]
```
`patch-file.yaml`:
```
- op: add
  path: /machine/install/extraKernelArgs
  value:
    - talos.logging.kernel=${udpAddress}
```

Notice it's using list of strings, so don't forget the `|-` if it's a
multiline patch.