feat: use HTTP POST request to generate schematic id
Because the image-factory server needs to know the requested schematic
id in order to be able to provide the image, we should do a HTTP POST
request to the server.

This will impact on Talhelper needing internet connection to work, and
also mean it will prone to failure in case the server is down or we can
not establish connection with the image factory server.

So there's `--offline-mode` flag (disabled by default) added to some
commands to overcome this.

In order to support possibility of selfhosted image-factory, there's a
new `imageFactory` object for `talconfig.yaml`. For example you can use
your selfhosted "http://my-factory.com" like so:

```
imageFactory:
  registryURL: my-factory.com
  schematicEndpoint: /schematics
  protocol: http
  installURLTmpl: {{.RegistryURL}}/installer/{{.ID}}:{{.Version}}
```