feat(config)!: `controlPlane` and `worker` is now type of `NodeConfigs`
This will allow all `NodeConfigs` fields to be applicable on node group
(`controlPlane` or `worker` depending on node type) or per node level.

The config will be "merged" with everything defined on per node level
take precedence, except for `patches` and `extraManifests` to keep the
behavior before this commit. Patches and extaManifests defined on both
node and node group level will be appended instead.

There are 2 new node options to change the behavior of `patches` and
`extraManifests` when defined at both the node and node group level.
Which are: `overridePatches` and `overrideExtraManifests`. By setting
them to `true` will make the `patches` and `extraManifests` defined in
node level to override the ones defined in node group level.

The config validation is now done after the "merging" is done for each
node. So, i.e when you define `schematic` incorrectly in `controlPlane`
struct, you might see the validation says you have incorrect
`nodes[0].schematic` instead assuming `nodes[0]` is a controlPlane node.

BREAKING CHANGE: This commit also remove all the deprecated fields which
are:
* `talosImageURL`
* all `inlinePatch`
* all `configPatches`
* `nodes[].extensions`