feat: bump the go module to `v3`
The only reason why v3 is created is to somehow fix the mistake I made
when bumping the project to v2. Apparently, having a `go` project
updated to v2 beyond will need some manual interventions that I didn't
know about.

The consequence is that the `go` module of this project after v2.0.0 tag
can't be imported from outside of this project because go require v2 to
be a separate module from v0 and v1 (https://go.dev/blog/v2-go-modules).

This commit will just bump the project to v3 because I've made too many
releases in v2 and fixing it will require a lot of headache.

Signed-off-by: budimanjojo <budimanjojo@gmail.com>